More improve the recovery of panic
Processing of some Result* moved to controller.Render* corresponding in
order to get the useful runtime stack when it panic.
The controller and method information was not contained in runtime
stack when it panic in Result.Proc because controller.Render* and
Result.Proc are run separately until now.
Originally panic in Result was not able to recover.